requirements.txt
Django>=5.2.8
django-simple-captcha>=0.6.0
Pillow>=10.0.0

